Generalize tag component

+++ src/components/UI/Tag/Tag.js
@@ -4,11 +4,17 @@ import { MdClear } from 'react-icons/md';
 
 import classes from './Tag.module.css';
 
-const tag = ({ children, onDelete }) =>
-  <span className={classes.Tag}>
+const tag = ({ children, onDelete, onClick }) => {
+  let deleteIcon="";
+  if (onDelete) {
+    deleteIcon = <MdClear onClick={() => onDelete()} />;
+  }
+
+  return <span className={classes.Tag} onClick={() => onClick()}>
     <span>{children}</span>
-    <MdClear onClick={() => onDelete()} />
+    {deleteIcon}
   </span>
+}
 
 
 export default tag;
